Use a middleware to make sure exceptions are shown
This also displays the queries that were executed and total time. This
is better than duplicating such code in each view.
Remove debug.html which was meant to play the same role, but is not
appropriate when using AJAX since it would only display the queries used
to display base.html, not the various AJAX blocks