Add support pyinstaller
A new spec fiel geneaprov.spec has been added. It can be used with
pyinstaller to provide a standalone package that contains everything
a user needs to run geneaprove, on Windows, OSX or linux.

Renames the database, so that the development one is not embedded
in the pyinstaller package.

Move initial_json.json so that it is available in that package.

Compiled resources are now copied to static/ rather than dist/ for
better compatibility with standard django layout and pyinstaller