Import FAM.SOUR
This is a source associated with the family, unclear which event is
related. For now, we create a new dummy event MARR so that the family
exists, though there's no date associated.

Duplicate events when there are multiple sources in gedcom

If an event ("BIRT") has multiple sources, we used to end up with
something like:
    P1 -- p2e (source1) ---> event
    P2 -- p2e (source2) --/
Thus when looking at the details on the persona page, we are seeing
two blocks for births (one for p1, one for p2), and when we expand
the event details, we see P1 and P2 in both cases.

Instead, we now create two events, so we end up with
    P1 -- p2e (source1) ---> event1
    P2 -- p2e (source2) ---> event2
We still end up with two blocks in the persona page (one for p1,
one for p2), but now expanding the details shows either P1 or P2,
not both.

We still need to improve things so that when P1 is marked "same as"
P2, there should be a way to mark event1 "same as" event2. Or
perhaps an assertion-to-assertion conclusion, so that we end up with

    P1 --p2e1 (source1)--> event1
    P2 --p2e2 (source2)--> event2
    P1 --p2g ---> group1 (all persons sameAs P1 and P2)
    P2 --p2g ---> group1
    group1 --g2p--> P3   (the merged persona)
    P3 --p2e3 (via assert2assert) -> event3

because p2e1 and p2e2 are basis for a higher-level assertion, they
should be filtered out, and we do not see event1 or event2 on the
persona page, and thus only event3 shows.