Minor improvements to PersonaList, PlaceList and SourceList
We use now react-window, instead of our custom SmartTable. This
component seems kind of standard, and means we have less complex code to
maintain, and perhaps performance will be even better.

Rewrite persona list using hooks instead of class component. In
particular, helped detect a bug where the list of personas wasn't
refreshed when changing the color theme.

Add new hooks to automatically compute the size of a component (instead
of hard-coding the width to 900px in javascript, it is now controlled
via CSS, which will make it possible to adapt to different window sizes.

Use debounce to filter the list of persons. We no longer re-filter on
every keypress, only after a small delay, which means the filter is
more efficient with very long lists.

Removing SmartTable and replacing with react-window resulted in the
following change in sizes:
     145.76 KB  build/static/js/1.64edc1e1.chunk.js
      95.78 KB   build/static/css/1.f81643f1.chunk.css
      34.49 KB   build/static/js/main.e3ee90a1.chunk.js
      2.6 KB     build/static/css/main.d1418694.chunk.css
      763 B      build/static/js/runtime~main.229c360f.js
to:
     139.75 KB  build/static/js/1.d6909d0c.chunk.js
     95.78 KB   build/static/css/1.22f30caa.chunk.css
     32.88 KB   build/static/js/main.e9a44e1e.chunk.js
     2.12 KB    build/static/css/main.e372ca90.chunk.css
     763 B      build/static/js/runtime~main.229c360f.js