Initial support for dark mode
Currently, only Safari 12.1 supports the requires CSS attributes.
Firefox 67 and Chrome 76 will shortly have them.

Integration is not very good yet, since we need to override semantic-ui
CSS in a number of places, so it would be better if semantic-ui itself
supported it.

Also some work to be done with the default color schemes used for
pedigree and fanchart, as some of them might use colors that do not look
good in dark mode.