Remove Persona.display_name field
This was set from the imported gedcom, so in the case of Gramps
means we were display names as "John /Smith/". Instead, it would
be better for the GUI to format the name from the name parts
that are stored in the database, so we can have various formatting
rules (use bold, given name first,...)
This is WIP, and queries have not been adjusted yet.